Исправил определение пути к файлу VERSION

+++ install/deb-build.sh
@@ -17,10 +17,20 @@ cp -r $BINPATH/* $TMPDIR/bin
 cp -r ../install/builders/deb/* $TMPDIR/deb
 cp -r ../oscript-library/* $TMPDIR/lib
 
+VERSIONFILE=$BINPATH/VERSION
+if [ -f "$VERSIONFILE" ] ; then
+	rm $VERSIONFILE
+fi
+
 mono ${BINPATH}/oscript.exe | head -1 | \
 		grep -oE '([[:digit:]]+\.){2}[[:digit:]]+' \
-		> ${BINPATH}VERSION
+		> ${BINPATH}/VERSION
 
+if [ ! -f "$VERSIONFILE" ] ; then
+	echo "No version file created"
+	exit 1
+fi
+		
 docker build -t onescript:deb ${PWD}/builders/deb/
 docker run --rm -v ${TMPDIR}:/media onescript:deb
